Only three remaining tests, possibly some hard multi root remaining though
