+++ test/ly/stealth/xmlavro/ConverterTest.java
@@ -77,44 +77,21 @@ public class ConverterTest {
     }
 
     @Test
-    public void severalRoots() {
-        String xsd =
-                "<xs:schema xmlns:xs='http://www.w3.org/2001/XMLSchema'>" +
-                "   <xs:element name='i' type='xs:int'/>" +
-                "   <xs:element name='r'>" +
-                "     <xs:complexType>" +
-                "       <xs:sequence>" +
-                "         <xs:element name='s' type='xs:string'/>" +
-                "       </xs:sequence>" +
-                "     </xs:complexType>" +
-                "   </xs:element>" +
-                "</xs:schema>";
-
-        Schema schema = Converter.createSchema(xsd);
-        assertEquals(Schema.Type.RECORD, schema.getType());
-        assertTrue("Schema should have a valid name", schema.getName() != null && !schema.getName().isEmpty());
-        assertEquals(Source.DOCUMENT, schema.getProp(Source.SOURCE));
-        assertEquals(2, schema.getFields().size());
-
-        Schema.Field field0 = schema.getFields().get(0);
-        assertEquals("" + new Source("i"), field0.getProp(Source.SOURCE));
-        assertEquals(Schema.Type.UNION, field0.schema().getType());
-        assertEquals(Schema.Type.INT, field0.schema().getTypes().get(1).getType());
-        assertEquals(Schema.Type.NULL, field0.schema().getTypes().get(0).getType());
-
-        Schema.Field field1 = schema.getFields().get(1);
-        assertEquals("" + new Source("r"), field1.getProp(Source.SOURCE));
-        assertEquals(Schema.Type.UNION, field1.schema().getType());
-        assertEquals(Schema.Type.RECORD, field1.schema().getTypes().get(1).getType());
-        assertEquals(Schema.Type.NULL, field1.schema().getTypes().get(0).getType());
+    public void severalRootsOne() {
+        Schema schema = Converter.createSchema(TestData.severalRoots.xsd);
 
         String xml = "<i>5</i>";
         GenericData.Record record = Converter.createDatum(schema, xml);
         assertEquals(null, record.get("r"));
         assertEquals(5, record.get("i"));
+    }
 
-        xml = "<r><s>s</s></r>";
-        record = Converter.createDatum(schema, xml);
+    @Test
+    public void severalRootsTwo() {
+        Schema schema = Converter.createSchema(TestData.severalRoots.xsd);
+
+        String xml = "<r><s>s</s></r>";
+        GenericData.Record record = Converter.createDatum(schema, xml);
         GenericData.Record subRecord = (GenericData.Record) record.get("r");
         assertEquals("s", subRecord.get("s"));
     }
@@ -217,27 +194,6 @@ public class ConverterTest {
     }
 
     @Test
-    public void severalWildcards() {
-        String xsd =
-                "<xs:schema xmlns:xs='http://www.w3.org/2001/XMLSchema'>" +
-                "  <xs:element name='root'>" +
-                "    <xs:complexType>" +
-                "      <xs:sequence>" +
-                "        <xs:any/>" +
-                "        <xs:any/>" +
-                "      </xs:sequence>" +
-                "    </xs:complexType>" +
-                "  </xs:element>" +
-                "</xs:schema>";
-
-        Schema schema = Converter.createSchema(xsd);
-        assertEquals(1, schema.getFields().size());
-
-        Schema.Field field = schema.getField(Source.WILDCARD);
-        assertEquals(null, field.getProp(Source.SOURCE));
-    }
-
-    @Test
     public void optionalElementValues() {
         Schema schema = Converter.createSchema(TestData.optionalElementValues.xsd);
         assertEquals(2, schema.getFields().size());
@@ -277,60 +233,19 @@ public class ConverterTest {
     }
 
     @Test
-    public void choiceElements() {
-        String xsd =
-                "<xs:schema xmlns:xs='http://www.w3.org/2001/XMLSchema'>" +
-                "  <xs:element name='root'>" +
-                "    <xs:complexType>" +
-                "      <xs:choice>" +
-                "        <xs:element name='s' type='xs:string'/>" +
-                "        <xs:element name='i' type='xs:int'/>" +
-                "      </xs:choice>" +
-                "    </xs:complexType>" +
-                "  </xs:element>" +
-                "</xs:schema>";
-
-        Schema schema = Converter.createSchema(xsd);
-        assertEquals(Schema.Type.RECORD, schema.getType());
-        assertEquals(2, schema.getFields().size());
-
-        Schema.Field sField = schema.getField("s");
-        assertEquals(Schema.Type.UNION, sField.schema().getType());
-        assertEquals(
-                Arrays.asList(Schema.create(Schema.Type.NULL), Schema.create(Schema.Type.STRING)),
-                sField.schema().getTypes()
-        );
-
-        Schema.Field iField = schema.getField("i");
-        assertEquals(Schema.Type.UNION, iField.schema().getType());
-        assertEquals(Arrays.asList(Schema.create(Schema.Type.NULL), Schema.create(Schema.Type.INT)), iField.schema().getTypes());
-
+    public void choiceElementsOne() {
+        Schema schema = Converter.createSchema(TestData.choiceElements.xsd);
         String xml = "<root><s>s</s></root>";
         GenericData.Record record = Converter.createDatum(schema, xml);
         assertEquals("s", record.get("s"));
-
-        xml = "<root><i>1</i></root>";
-        record = Converter.createDatum(schema, xml);
-        assertEquals(1, record.get("i"));
     }
 
     @Test
-    public void arrayOfUnboundedChoiceElements() {
-      String xsd = "<xs:schema xmlns:xs='http://www.w3.org/2001/XMLSchema'>" +
-              "  <xs:element name='root'>" +
-              "    <xs:complexType>" +
-              "      <xs:choice maxOccurs='unbounded'>" +
-              "        <xs:element name='s' type='xs:string'/>" +
-              "        <xs:element name='i' type='xs:int'/>" +
-              "      </xs:choice>" +
-              "    </xs:complexType>" +
-              "  </xs:element>" +
-              "</xs:schema>";
-
-      Schema schema = Converter.createSchema(xsd);
-      assertEquals(Schema.Type.ARRAY, schema.getType());
-      final Schema elementType = schema.getElementType();
-      assertEquals(Schema.Type.RECORD, elementType.getType());
+    public void choiceElementsTwo() {
+        Schema schema = Converter.createSchema(TestData.choiceElements.xsd);
+        String xml = "<root><i>1</i></root>";
+        GenericData.Record  record = Converter.createDatum(schema, xml);
+        assertEquals(1, record.get("i"));
     }
 
     @Test
